backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
python-multipart
bcrypt
email-validator
python-dotenv
pydantic-settings